added an example for arduino development board
Arduino listens to its serial input and extracts specific text commands
from it. Once a proper command is extracted it tells a servo motor
attached to it to turn to a certain degree. A webcam attached to the
servo provides its output to the Rift. Thus, we have a webcam controlled
by the Rift.